requirements.txt
Django>=4.1.3,<4.2.0
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.5,<3.0.0
Pillow>=9.3.0,<9.4.0

flake8>=6.0.0,<6.1.0